+ Add a few more service functions that wrap basic cuda commands   (e.g.: MemAlloc and MemFree) and also do error checking. + Update GPU_Module so it can find and return a handle to module   functions. + Update unit tests to demonstrate use of both the wrapped and   unwrapped cuda calls.

+++ src/device/test/vector_add.cu
@@ -1,7 +1,7 @@
 extern "C"
 {
     __global__ void
-    vector_add(float const *A, float const *B, float *C, int const N)
+    vector_add(double const *A, double const *B, double *C, int const N)
     {
         int i = blockDim.x * blockIdx.x + threadIdx.x; 
         // if(i%512==0)
